Set theory 2 (#133)
* Rename functions for consistency

* Pair deconstruction functions

* Sets of functions, properties about functions, application; unimplemented inverse functions

* Documentation consistency

* Change partial orders to use pair; increasing functions and isomorphisms

* Split set theory files into a better strucutre

* Deleted old set theory files

* Add top and bot defs to running theory
